Fix deps with brave_wallet constants
