+++ browser/BUILD.gn
@@ -201,6 +201,7 @@ source_set("browser_process") {
   if (brave_wallet_enabled) {
     deps += [
       "//brave/components/brave_wallet/browser",
+      "//brave/components/brave_wallet/common",
       "brave_wallet",
     ]
   }
