+++ browser/extensions/BUILD.gn
@@ -49,16 +49,6 @@ source_set("extensions") {
     "updater/brave_update_client_config.h",
   ]
 
-  if (brave_wallet_enabled) {
-    sources += [
-      "api/brave_wallet_api.cc",
-      "api/brave_wallet_api.h",
-      "brave_wallet_util.cc",
-      "brave_wallet_util.h"
-    ]
-    configs += [ ":infura_config" ]
-  }
-
   if (binance_enabled) {
     sources += [
       "api/binance_api.cc",
@@ -110,6 +100,20 @@ source_set("extensions") {
   if (ntp_widget_utils_enabled) {
     deps += [
       "//brave/components/ntp_widget_utils/browser",
+    ]
+  }
+
+  if (brave_wallet_enabled) {
+    sources += [
+      "api/brave_wallet_api.cc",
+      "api/brave_wallet_api.h",
+      "brave_wallet_util.cc",
+      "brave_wallet_util.h"
+    ]
+    configs += [ ":infura_config" ]
+    deps += [
+      "//brave/components/brave_wallet/browser",
+      "//brave/components/brave_wallet/common",
     ]
   }
