+++ browser/extensions/api/brave_wallet_api.cc
@@ -13,9 +13,10 @@
 #include "brave/browser/extensions/brave_wallet_util.h"
 #include "brave/browser/infobars/crypto_wallets_infobar_delegate.h"
 #include "brave/browser/profiles/profile_util.h"
-#include "brave/common/brave_wallet_constants.h"
 #include "brave/common/extensions/api/brave_wallet.h"
 #include "brave/common/pref_names.h"
+#include "brave/components/brave_wallet/common/brave_wallet_constants.h"
+#include "brave/components/brave_wallet/common/brave_wallet_pref_names.h"
 #include "brave/grit/brave_generated_resources.h"
 #include "chrome/browser/extensions/api/tabs/tabs_constants.h"
 #include "chrome/browser/extensions/extension_tab_util.h"
