+++ browser/extensions/api/brave_wallet_api_browsertest.cc
@@ -7,8 +7,9 @@
 #include "base/scoped_observer.h"
 #include "brave/browser/infobars/crypto_wallets_infobar_delegate.h"
 #include "brave/common/brave_paths.h"
-#include "brave/common/brave_wallet_constants.h"
 #include "brave/common/pref_names.h"
+#include "brave/components/brave_wallet/common/brave_wallet_constants.h"
+#include "brave/components/brave_wallet/common/brave_wallet_pref_names.h"
 #include "chrome/browser/infobars/infobar_service.h"
 #include "chrome/browser/profiles/profile.h"
 #include "chrome/browser/ui/browser.h"
