+++ browser/infobars/BUILD.gn
@@ -36,7 +36,8 @@ source_set("infobars") {
     deps += [
       "//brave/browser/brave_wallet:brave_wallet",
       "//brave/common:pref_names",
-      "//brave/components/brave_wallet/browser:browser",
+      "//brave/components/brave_wallet/browser",
+      "//brave/components/brave_wallet/common",
       "//components/user_prefs:user_prefs",
       "//extensions/common:common_constants",
     ]
