+++ browser/ipfs/ipfs_tab_helper_browsertest.cc
@@ -7,7 +7,6 @@
 #include "base/scoped_observer.h"
 #include "brave/browser/ipfs/ipfs_tab_helper.h"
 #include "brave/common/brave_paths.h"
-#include "brave/common/brave_wallet_constants.h"
 #include "brave/common/pref_names.h"
 #include "brave/components/ipfs/browser/features.h"
 #include "brave/components/ipfs/common/ipfs_constants.h"
