+++ chromium_src/extensions/common/constants.cc
@@ -18,25 +18,6 @@ const char brave_webtorrent_extension_id[] = "lgjmpdmojkpocjcopdikifhejkkjglho";
 const char brave_webtorrent_extension_filename[] = "brave_webtorrent.html";
 const char brave_webtorrent_extension_filename2[] = "brave_webtorrent2.html";
 
-const char ethereum_remote_client_extension_id[] =
-    "odbfpeeihdkbihmopkbjmoonfanlbfcl";
-const char ethereum_remote_client_extension_name[] = "Crypto Wallets";
-const char ethereum_remote_client_extension_public_key[] =
-    "MIIBIjANBgkqhkiG9w0BAQEFAAOCAQ8AMIIBCgKCAQEAl7ZLkqbq8pHRcYANANXmhGKZKHsru"
-    "BqS0ssf8aI5R5E5FBaWui73FPWMj0g4eggrJeSfF/x/nIIq3Z24Joq1ULuWtheQjbOwqj0yL2"
-    "ZaxMA6rc5cW2OmGZ0aNXMFclusXftJmwpb/pztHzw55yv8BfJr873HOxtlW2MP1VvSFQjigsb"
-    "JedlkoS4SKC0U3j8/mjAYR0+lpmBtPitObnYnM5qmtrlg2hgm+sqAon9WKj9nDIjPpXJaCs7k"
-    "pVl7oQOoYpv47ZT1tnu0o03AL3RZbxPU1N04X3JN6nC+g4gaCgiLkZ+dD79yIt7Il0+vSKuJ+"
-    "EGySVmv6czOI5eJYezWXwIDAQAB";
-const char ethereum_remote_client_base_url[] =
-    "chrome-extension://odbfpeeihdkbihmopkbjmoonfanlbfcl/home.html";
-const char ethereum_remote_client_phishing_url[] =
-    "chrome-extension://odbfpeeihdkbihmopkbjmoonfanlbfcl/phishing.html";
-const char ethereum_remote_client_ens_redirect_url[] =
-    "chrome-extension://odbfpeeihdkbihmopkbjmoonfanlbfcl/loading.html";
-const char ethereum_remote_client_host[] = "wallet";
-const char metamask_extension_id[] = "nkbihfbeogaeaoehlefnkodbefgpgknn";
-
 const char ipfs_companion_extension_id[] = "nibjojkomfdiaoajekhjakgkdhaomnch";
 const char ipfs_companion_extension_name[] = "IPFS Companion";
 const char ipfs_companion_extension_public_key[] =
