+++ chromium_src/extensions/common/constants.h
@@ -21,15 +21,6 @@ extern const char brave_webtorrent_extension_id[];
 extern const char brave_webtorrent_extension_filename[];
 extern const char brave_webtorrent_extension_filename2[];
 
-extern const char ethereum_remote_client_extension_id[];
-extern const char ethereum_remote_client_extension_name[];
-extern const char ethereum_remote_client_extension_public_key[];
-extern const char ethereum_remote_client_base_url[];
-extern const char ethereum_remote_client_phishing_url[];
-extern const char ethereum_remote_client_ens_redirect_url[];
-extern const char ethereum_remote_client_host[];
-extern const char metamask_extension_id[];
-
 extern const char ipfs_companion_extension_id[];
 extern const char ipfs_companion_extension_name[];
 extern const char ipfs_companion_extension_public_key[];
