--- common/brave_wallet_constants.h
@@ -1,16 +1,0 @@
-/* Copyright (c) 2020 The Brave Authors. All rights reserved.
- * This Source Code Form is subject to the terms of the Mozilla Public
- * License, v. 2.0. If a copy of the MPL was not distributed with this file,
- * You can obtain one at http://mozilla.org/MPL/2.0/. */
-
-#ifndef BRAVE_COMMON_BRAVE_WALLET_CONSTANTS_H_
-#define BRAVE_COMMON_BRAVE_WALLET_CONSTANTS_H_
-
-enum class BraveWalletWeb3ProviderTypes {
-  ASK,
-  NONE,
-  CRYPTO_WALLETS,
-  METAMASK
-};
-
-#endif  // BRAVE_COMMON_BRAVE_WALLET_CONSTANTS_H_
