+++ common/pref_names.cc
@@ -65,14 +65,6 @@ const char kNewTabPageShowGemini[] = "brave.new_tab_page.show_gemini";
 const char kNewTabPageShowTogether[] = "brave.new_tab_page.show_together";
 const char kNewTabPageShowAddCard[] = "brave.new_tab_page.show_addcard";
 const char kBraveEnabledMediaRouter[] = "brave.enable_media_router";
-const char kBraveWalletAES256GCMSivNonce[] =
-    "brave.wallet.aes_256_gcm_siv_nonce";
-const char kBraveWalletEncryptedSeed[] = "brave.wallet.encrypted_seed";
-const char kBraveWalletEnabledDeprecated[] = "brave.wallet.enabled";
-const char kBraveWalletPrefVersion[] = "brave.wallet.pref_version";
-const char kBraveWalletWeb3Provider[] = "brave.wallet.web3_provider";
-const char kLoadCryptoWalletsOnStartup[] = "brave.wallet.load_on_startup";
-const char kOptedIntoCryptoWallets[] = "brave.wallet.opted_in";
 const char kBinanceAccessToken[] = "brave.binance.access_token";
 const char kBinanceRefreshToken[] = "brave.binance.refresh_token";
 const char kAlwaysShowBookmarkBarOnNTP[] =
