+++ components/brave_wallet/browser/brave_wallet_service.cc
@@ -13,9 +13,9 @@
 #include "base/files/file_util.h"
 #include "base/task/post_task.h"
 #include "base/task_runner_util.h"
-#include "brave/common/brave_wallet_constants.h"
-#include "brave/common/pref_names.h"
 #include "brave/components/brave_wallet/browser/brave_wallet_delegate.h"
+#include "brave/components/brave_wallet/common/brave_wallet_constants.h"
+#include "brave/components/brave_wallet/common/brave_wallet_pref_names.h"
 #include "components/user_prefs/user_prefs.h"
 #include "content/public/browser/browser_context.h"
 #include "crypto/aead.h"
