+++ components/brave_wallet/common/BUILD.gn
@@ -1,0 +1,12 @@
+import("//brave/components/brave_wallet/browser/buildflags/buildflags.gni")
+
+assert(brave_wallet_enabled)
+
+source_set("common") {
+    sources = [
+        "brave_wallet_constants.cc",
+        "brave_wallet_constants.h",
+        "brave_wallet_pref_names.cc",
+        "brave_wallet_pref_names.h",
+    ]
+}
