+++ components/brave_wallet/common/brave_wallet_constants.cc
@@ -1,0 +1,25 @@
+/* Copyright (c) 2020 The Brave Authors. All rights reserved.
+ * This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this file,
+ * You can obtain one at http://mozilla.org/MPL/2.0/. */
+
+#include "brave/components/brave_wallet/common/brave_wallet_constants.h"
+
+const char ethereum_remote_client_extension_id[] =
+    "odbfpeeihdkbihmopkbjmoonfanlbfcl";
+const char ethereum_remote_client_extension_name[] = "Crypto Wallets";
+const char ethereum_remote_client_extension_public_key[] =
+    "MIIBIjANBgkqhkiG9w0BAQEFAAOCAQ8AMIIBCgKCAQEAl7ZLkqbq8pHRcYANANXmhGKZKHsru"
+    "BqS0ssf8aI5R5E5FBaWui73FPWMj0g4eggrJeSfF/x/nIIq3Z24Joq1ULuWtheQjbOwqj0yL2"
+    "ZaxMA6rc5cW2OmGZ0aNXMFclusXftJmwpb/pztHzw55yv8BfJr873HOxtlW2MP1VvSFQjigsb"
+    "JedlkoS4SKC0U3j8/mjAYR0+lpmBtPitObnYnM5qmtrlg2hgm+sqAon9WKj9nDIjPpXJaCs7k"
+    "pVl7oQOoYpv47ZT1tnu0o03AL3RZbxPU1N04X3JN6nC+g4gaCgiLkZ+dD79yIt7Il0+vSKuJ+"
+    "EGySVmv6czOI5eJYezWXwIDAQAB";
+const char ethereum_remote_client_base_url[] =
+    "chrome-extension://odbfpeeihdkbihmopkbjmoonfanlbfcl/home.html";
+const char ethereum_remote_client_phishing_url[] =
+    "chrome-extension://odbfpeeihdkbihmopkbjmoonfanlbfcl/phishing.html";
+const char ethereum_remote_client_ens_redirect_url[] =
+    "chrome-extension://odbfpeeihdkbihmopkbjmoonfanlbfcl/loading.html";
+const char ethereum_remote_client_host[] = "wallet";
+const char metamask_extension_id[] = "nkbihfbeogaeaoehlefnkodbefgpgknn";
