+++ components/brave_wallet/common/brave_wallet_constants.h
@@ -1,0 +1,25 @@
+/* Copyright (c) 2020 The Brave Authors. All rights reserved.
+ * This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this file,
+ * You can obtain one at http://mozilla.org/MPL/2.0/. */
+
+#ifndef BRAVE_COMPONENTS_BRAVE_WALLET_COMMON_BRAVE_WALLET_CONSTANTS_H_
+#define BRAVE_COMPONENTS_BRAVE_WALLET_COMMON_BRAVE_WALLET_CONSTANTS_H_
+
+enum class BraveWalletWeb3ProviderTypes {
+  ASK,
+  NONE,
+  CRYPTO_WALLETS,
+  METAMASK
+};
+
+extern const char ethereum_remote_client_extension_id[];
+extern const char ethereum_remote_client_extension_name[];
+extern const char ethereum_remote_client_extension_public_key[];
+extern const char ethereum_remote_client_base_url[];
+extern const char ethereum_remote_client_phishing_url[];
+extern const char ethereum_remote_client_ens_redirect_url[];
+extern const char ethereum_remote_client_host[];
+extern const char metamask_extension_id[];
+
+#endif  // BRAVE_COMPONENTS_BRAVE_WALLET_COMMON_BRAVE_WALLET_CONSTANTS_H_
