+++ components/brave_wallet/common/brave_wallet_pref_names.cc
@@ -1,0 +1,15 @@
+/* Copyright (c) 2020 The Brave Authors. All rights reserved.
+ * This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this file,
+ * You can obtain one at http://mozilla.org/MPL/2.0/. */
+
+#include "brave/components/brave_wallet/common/brave_wallet_pref_names.h"
+
+const char kBraveWalletAES256GCMSivNonce[] =
+    "brave.wallet.aes_256_gcm_siv_nonce";
+const char kBraveWalletEncryptedSeed[] = "brave.wallet.encrypted_seed";
+const char kBraveWalletEnabledDeprecated[] = "brave.wallet.enabled";
+const char kBraveWalletPrefVersion[] = "brave.wallet.pref_version";
+const char kBraveWalletWeb3Provider[] = "brave.wallet.web3_provider";
+const char kLoadCryptoWalletsOnStartup[] = "brave.wallet.load_on_startup";
+const char kOptedIntoCryptoWallets[] = "brave.wallet.opted_in";
